extracted logic responsible for retrieving ambulance id and dispacth data from combo box component and moved it to patient fields component
